Fix unit test PHPCS violations for alerts collaborators.
Move WP_Query and WP_Roles stubs into dedicated files, add phpcs:ignore
for intentional $GLOBALS overrides in menu tests, and fix foreach closing
comment formatting so lint-tests passes in CI.
